+++ src/main/java/ojosama/talkak/video/service/VideoService.java
@@ -6,7 +6,7 @@ import ojosama.talkak.video.dto.YoutubeUrlValidationAPIResponse;
 import ojosama.talkak.video.dto.YoutubeUrlValidationRequest;
 import ojosama.talkak.video.dto.YoutubeUrlValidationResponse;
 import ojosama.talkak.video.util.IdExtractor;
-import ojosama.talkak.video.util.WebClientUtil;
+import ojosama.talkak.common.util.WebClientUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
